+++ ElvUI_LuckyoneUI/Core/Install.lua
@@ -9,7 +9,7 @@ local function InstallComplete()
 	ReloadUI()
 end
 
--- Installer Table
+-- Installer table
 L1UI.InstallerData = {
 	Title = format('|cff4beb2c%s %s|r', L1UI.Name, L["Installation"]),
 	Name = L1UI.Name,