+++ ElvUI_LuckyoneUI/Profiles/AddOnSkins.lua
@@ -9,19 +9,20 @@ function L1UI:Setup_AddOnSkins(noPrint)
 
 	local AS = unpack(AddOnSkins)
 
-	AS.data:SetProfile("Luckyone")
+	-- Create and set a new profile called Luckyone
+	AS.data:SetProfile('Luckyone')
 
-	AS.db["DBMFont"] = "Expressway"
-	AS.db["Shadows"] = false
-	AS.db["SkinDebug"] = true
+	AS.db.DBMFont = 'Expressway'
+	AS.db.Shadows = false
+	AS.db.SkinDebug = true
 
-	AS.db["EmbedBackdrop"] = false
-	AS.db["EmbedBackdropTransparent"] = false
-	AS.db["EmbedOoC"] = false
-	AS.db["EmbedRightChat"] = false
-	AS.db["EmbedSystem"] = false
-	AS.db["EmbedSystemDual"] = false
-	AS.db["EmbedSystemMessage"] = false
+	AS.db.EmbedBackdrop = false
+	AS.db.EmbedBackdropTransparent = false
+	AS.db.EmbedOoC = false
+	AS.db.EmbedRightChat = false
+	AS.db.EmbedSystem = false
+	AS.db.EmbedSystemDual = false
+	AS.db.EmbedSystemMessage = false
 
 	if not noPrint then
 		L1UI:Print(L["AddOnSkins profile has been set."])