+++ CHANGELOG.md
@@ -1,8 +1,18 @@
-### Version 1.53 [In Development]
+### Version 1.54 [In Development]
 **All WoW Versions:**  
-More updates for the Archive and Database  
-Reworked the configuration tab slightly  
+Minor tweaks for the initial ElvUI layout installer steps  
+**Retail:**  
+Updated Details profile  
+Updated ElvUI StyleFilters for Season 4  
+___
+### Version 1.53 [30-07-2022]
+**All WoW Versions:**  
+More updates for the archive and database  
+Reworked the configuration panel slightly  
 **Wrath Classic:**  
+Added Moonkin WeakAuras pack to the config  
+CVars installer step will now enable the Equipment Manager  
+CVars installer step will now enable Talent Preview mode  
 Fixed an error in the Plater installer step  
 ___
 ### Version 1.52 [24-07-2022]